@@ -23,19 +23,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.12",
-    "@angular/common": "^9.1.12",
-    "@angular/compiler": "^9.1.12",
-    "@angular/core": "^9.1.12",
-    "@angular/forms": "^9.1.12",
-    "@angular/localize": "^9.1.12",
-    "@angular/platform-browser": "^9.1.12",
-    "@angular/platform-browser-dynamic": "^9.1.12",
-    "@angular/platform-server": "^9.1.12",
-    "@angular/router": "^9.1.12",
+    "@angular/animations": "^10.1.0",
+    "@angular/common": "^10.1.0",
+    "@angular/compiler": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/forms": "^10.1.0",
+    "@angular/localize": "^10.1.0",
+    "@angular/platform-browser": "^10.1.0",
+    "@angular/platform-browser-dynamic": "^10.1.0",
+    "@angular/platform-server": "^10.1.0",
+    "@angular/router": "^10.1.0",
     "@hmcts/frontend": "^0.0.41-alpha",
     "@microsoft/signalr": "^3.1.4",
-    "@ng-select/ng-select": "^4.0.4",
+    "@ng-select/ng-select": "^5.0.1",
     "@popperjs/core": "^2.4.3",
     "adal-angular4": "^3.0.16",
     "applicationinsights-js": "^1.0.21",
@@ -47,22 +47,22 @@
     "guid-typescript": "^1.0.9",
     "moment": "^2.27.0",
     "ng-bullet": "^1.0.3",
-    "ng2-charts": "^2.3.2",
+    "ng2-charts": "^2.4.0",
     "ngx-clipboard": "^12.3.1",
-    "ngx-device-detector": "^1.4.4",
+    "ngx-device-detector": "^2.0.0",
     "ngx-moment": "^3.5.0",
     "run-script-os": "^1.1.1",
     "rxjs": "^6.5.5",
     "snyk": "^1.352.1",
     "webrtc-adapter": "^7.6.3",
     "zone.js": "~0.10.3",
-    "tslib": "^1.10.0"
+    "tslib": "^2.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.12",
-    "@angular/cli": "^9.1.12",
-    "@angular/compiler-cli": "^9.1.12",
-    "@angular/language-service": "^9.1.12",
+    "@angular-devkit/build-angular": "~0.1001.0",
+    "@angular/cli": "^10.1.0",
+    "@angular/compiler-cli": "^10.1.0",
+    "@angular/language-service": "^10.1.0",
     "@testpossessed/ts-data-builder": "^1.0.1",
     "@types/applicationinsights-js": "^1.0.9",
     "@types/chart.js": "^2.9.22",
@@ -70,27 +70,27 @@
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "~12.12.47",
     "@types/webrtc": "^0.0.26",
-    "codelyzer": "^5.2.2",
+    "codelyzer": "^6.0.0",
     "install-peers": "^1.0.3",
     "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.2",
-    "karma": "^5.0.9",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.3",
-    "karma-jasmine": "~3.3.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "karma-junit-reporter": "^2.0.1",
     "node-sass": "^4.13.1",
     "nswag": "^13.6.2",
     "prettier": "^2.0.5",
     "puppeteer": "^3.3.0",
-    "typescript": "~3.8.3"
+    "typescript": "~4.0.2"
   },
   "optionalDependencies": {
     "fsevents": "^2.1.3",
-    "protractor": "~5.4.0",
+    "protractor": "~7.0.0",
     "ts-node": "~5.0.1",
-    "tslint": "~5.19.0"
+    "tslint": "~6.1.0"
   },
   "snyk": true
 }
