@@ -7311,41 +7311,41 @@ namespace VideoWeb.Services.Video
         [System.Runtime.Serialization.EnumMember(Value = @"Help")]
         Help = 4,
     
-        [System.Runtime.Serialization.EnumMember(Value = @"Pause")]
-        Pause = 5,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"Close")]
-        Close = 6,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"Leave")]
-        Leave = 7,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"Consultation")]
-        Consultation = 8,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"MediaPermissionDenied")]
-        MediaPermissionDenied = 9,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"ParticipantJoining")]
-        ParticipantJoining = 10,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"SelfTestFailed")]
-        SelfTestFailed = 11,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"Suspend")]
-        Suspend = 12,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"VhoCall")]
-        VhoCall = 13,
-    
-        [System.Runtime.Serialization.EnumMember(Value = @"ParticipantNotSignedIn")]
-        ParticipantNotSignedIn = 14,
-    
         [System.Runtime.Serialization.EnumMember(Value = @"Start")]
-        Start = 15,
+        Start = 5,
     
         [System.Runtime.Serialization.EnumMember(Value = @"CountdownFinished")]
-        CountdownFinished = 16,
+        CountdownFinished = 6,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"Pause")]
+        Pause = 7,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"Close")]
+        Close = 8,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"Leave")]
+        Leave = 9,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"Consultation")]
+        Consultation = 10,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"MediaPermissionDenied")]
+        MediaPermissionDenied = 11,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"ParticipantJoining")]
+        ParticipantJoining = 12,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"SelfTestFailed")]
+        SelfTestFailed = 13,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"Suspend")]
+        Suspend = 14,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"VhoCall")]
+        VhoCall = 15,
+    
+        [System.Runtime.Serialization.EnumMember(Value = @"ParticipantNotSignedIn")]
+        ParticipantNotSignedIn = 16,
     
         [System.Runtime.Serialization.EnumMember(Value = @"EndpointJoined")]
         EndpointJoined = 17,
