@@ -1,8 +1,9 @@
-import { HostListener } from '@angular/core';
+import { HostListener, Directive } from '@angular/core';
 import { ParticipantStatusUpdateService } from 'src/app/services/participant-status-update.service';
 import { EventType } from 'src/app/services/clients/api-client';
 import { Logger } from 'src/app/services/logging/logger-base';
 
+@Directive()
 export abstract class ParticipantStatusBase {
     constructor(protected participantStatusUpdateService: ParticipantStatusUpdateService, protected logger: Logger) {}
 
